Move useSidebarNodes hook to local folder
